+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0+1
+
+* Email added to pubspec.yaml
+
 ## 1.0.0 - 2019-01-03
 
 * Added `FutureWidget`, `RefreshWidget`, `SearchWidget`